requirements.txt
discord.py[voice]
yt-dlp
ffmpeg
python-dotenv